oldbackend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==11.2.1
opencv-python==4.8.1.78
numpy==1.26.0 